+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Release Notes for Campaign
 
 ## 2.0.0-beta.4 - Unreleased
+### Added
+- Added a "Default Notification Contacts" field to sendout settings. 
+
 ### Changed
+- Replaced the "Notification Email Address" field with a "Notification Contacts" field in sendouts. 
 - Removed segments from "Singular" sendout types.
 - Renamed the `maxSendFailsAllowed` config setting to `maxSendFailuresAllowed`.